refactor(api): convert ThumbHash backfill into v17 schema upgrade
The standalone backfill (run via `npm run backfill-thumbhashes`) is replaced
by a proper schema upgrade so the ThumbHash backfill runs automatically as
part of the upgrade chain instead of being a manual post-startup step.

- Add schemaUpgrade/v17.ts: backfills ThumbHash placeholders for pre-existing
  Post/Tag images (fetch smallest stored variant from S3, encode, re-save
  parent, re-denormalise imageData onto child Content parentImageData). Guards
  on schema version 15 or 16 (v16 is reserved by a parallel branch) and bumps
  to 17. Per-image fetch/encode failures are logged and skipped so a missing
  S3 object can't block startup; DB-level failures re-throw to leave the
  version un-bumped for retry, and the per-collection thumbHash skip keeps the
  retry idempotent.
- Wire v17 into db.upgrade.ts after v15.
- Remove the standalone entrypoint (main.ts branch + package.json script) and
  delete db/backfillThumbHashes.ts.
- README: bump baseline to 17 and document the v17 upgrade.
