+++ api/package.json
@@ -19,7 +19,6 @@
         "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
         "test:e2e": "jest --config ./test/jest-e2e.json",
         "seed": "nest start -- seed",
-        "backfill-thumbhashes": "nest start -- backfill-thumbhashes",
         "auth-setup": "ts-node scripts/add-auth-provider.ts",
         "typecheck": "tsc --noEmit"
     },
