+++ api/src/db/db.upgrade.spec.ts
@@ -30,6 +30,10 @@ jest.mock("./schemaUpgrade/v16", () => ({
     __esModule: true,
     default: jest.fn().mockResolvedValue(undefined),
 }));
+jest.mock("./schemaUpgrade/v17", () => ({
+    __esModule: true,
+    default: jest.fn().mockResolvedValue(undefined),
+}));
 
 import { upgradeDbSchema } from "./db.upgrade";
 import v9 from "./schemaUpgrade/v9";
@@ -40,6 +44,7 @@ import v13 from "./schemaUpgrade/v13";
 import v14 from "./schemaUpgrade/v14";
 import v15 from "./schemaUpgrade/v15";
 import v16 from "./schemaUpgrade/v16";
+import v17 from "./schemaUpgrade/v17";
 
 describe("upgradeDbSchema", () => {
     const mockDb = {} as any;
@@ -59,6 +64,7 @@ describe("upgradeDbSchema", () => {
         expect(v14).toHaveBeenCalledWith(mockDb);
         expect(v15).toHaveBeenCalledWith(mockDb);
         expect(v16).toHaveBeenCalledWith(mockDb);
+        expect(v17).toHaveBeenCalledWith(mockDb);
     });
 
     it("should re-throw error and log it when an upgrade function fails", async () => {
