+++ api/src/db/db.upgrade.ts
@@ -7,6 +7,7 @@ import v13 from "./schemaUpgrade/v13";
 import v14 from "./schemaUpgrade/v14";
 import v15 from "./schemaUpgrade/v15";
 import v16 from "./schemaUpgrade/v16";
+import v17 from "./schemaUpgrade/v17";
 
 /**
  * Upgrade the database schema
@@ -22,6 +23,7 @@ export async function upgradeDbSchema(db: DbService) {
         await v14(db);
         await v15(db);
         await v16(db);
+        await v17(db);
     } catch (error) {
         console.error("Database schema upgrade failed:", error);
         throw error; // Re-throw to prevent schema version from being updated
